add custom taxonomie on articles

+++ web/app/themes/custom/metaboxes/SponsoMetaBox.php
@@ -5,6 +5,7 @@ namespace App\Metaboxes;
 class SponsoMetaBox
 {
     const META_KEY = 'customsponso';
+    const NONCE = '_custom_sponso_nonce';
 
     public static function register() {
         add_action('add_meta_boxes', [self::class, 'add'], 10, 2);
@@ -21,6 +22,7 @@ class SponsoMetaBox
     public static function render($post)
     {
         $value = get_post_meta($post->ID, self::META_KEY, true);
+        wp_nonce_field(self::NONCE, self::NONCE);
         ?>
         <input type="hidden" value="0" name="<?= self::META_KEY ?>">
         <input type="checkbox" value="1" name="<?= self::META_KEY ?>" <?= checked($value, '1')?>>
@@ -30,7 +32,11 @@ class SponsoMetaBox
 
     public static function save(int $post)
     {
-        if (array_key_exists(self::META_KEY, $_POST) && current_user_can('publish_post', $post)) {
+        if (
+            array_key_exists(self::META_KEY, $_POST) &&
+            current_user_can('publish_post', $post) &&
+            wp_verify_nonce($_POST[self::NONCE], self::NONCE)
+            ) {
             if ($_POST[self::META_KEY] === '0') {
                 delete_post_meta($post, self::META_KEY);
             } else {
